Fix int again... and please code in python...

+++ examples/plotting/file/tap.py
@@ -13,9 +13,8 @@ y = yy.flatten()
 N = len(x)
 inds = [str(i) for i in np.arange(N)]
 radii = np.random.random(size=N)*0.4 + 1.7
-colors = [
-    "#%02x%02x%02x" % (r, g, 150) for int(r), int(g) in zip(np.floor(50+2*x), np.floor(30+2*y))
-]
+colors = ["#%02x%02x%02x" % (r, g, 150) for r, g in
+          zip(np.floor(50+2*x).astype(int), np.floor(30+2*y).astype(int))]
 
 foo = list(itertools.permutations("abcdef"))[:N]
 bar = np.random.normal(size=N)
